+++ __tests__/index.ts
@@ -1,3 +1,5 @@
+import { memcache } from '@pulumi/gcp'
+
 export * as biguery from './bigquery'
 export * as cloudfunctions from './cloudfunctions'
 export * as cloudfunctionsv2 from './cloudfunctionsv2'
@@ -6,6 +8,7 @@ export * as cloudrunv2 from './cloudrunv2'
 export * as compute from './compute'
 export * as container from './container'
 export * as dns from './dns'
+export * as memcache from './memcache'
 export * as pubsub from './pubsub'
 export * as redis from './redis'
 export * as sql from './sql'